implemented slider to cart
