+++ js/data/constants.js
@@ -6,3 +6,9 @@ export const productSliderContainer = document.querySelector(".popular_products_
 
 // Product page
 export const imageContainer = document.querySelector(".product-image-container");
+
+export const productTitleContainer = document.querySelector(".product_choises_headline");
+export const productDescriptionContainer = document.querySelector(".product_choises_list");
+export const productSizeContainer = document.querySelector(".size_option_container");
+export const productPriceContainer = document.querySelector(".price-container");
+export const productProductDetailsContainer = document.querySelector(".product_details");
