+++ js/pages/cart.js
@@ -1,0 +1,5 @@
+import { fetchJackets } from "../components/productSlider.js";
+// import { fetchJacket } from "../components/fetchJacket.js";
+
+fetchJackets();
+// fetchJacket();
